Move decimal accumulator to new file
